FinnhubProducer/requirements.txt
kafka-python-ng
websocket-client
python-dotenv
